rename ws hooks files
